requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
attrs==23.2.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
croniter==3.0.3
discord-py-interactions==5.12.1
discord-typings==0.9.0
discord.py==2.4.0
emoji==2.12.1
frozenlist==1.4.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
idna==3.7
multidict==6.0.5
numpy==2.1.0
oauthlib==3.2.2
pandas==2.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
yarl==1.9.4
